requirements.txt
# ---- Original ML stack (kept as-is except for upgrades noted below) ----
keras==2.8.0
keras_tuner==1.1.2
kfp==1.8.6
kfp_pipeline_spec==0.1.16
matplotlib==3.4.1
minio==5.0.10
numpy==1.24.0
pandas==1.4.2
plotly==5.10.0
scikit_learn==1.1.1
seaborn==0.11.2
tensorboard==2.8.0

# ---- Upgraded common dependencies (required by FastAPI) ----
openpyxl==3.1.5          # upgraded from 3.0.9
pydantic==2.11.7         # upgraded from 1.9.2
requests==2.32.4         # upgraded from 2.28.2
typer==0.16.0            # upgraded from 0.4.1

# ---- Additional FastAPI-only dependencies ----
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.6
botocore==1.40.6
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
et-xmlfile==2.0.0
fastapi==0.111.0
fastapi-cli==0.0.8
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
orjson==3.11.1
pyarrow==7.0.0           # originally in ML list
pyasn1==0.6.1
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pyyaml==6.0.2
redis==5.0.7
rich==14.1.0
rich-toolkit==0.14.9
rsa==4.9.1
s3transfer==0.13.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
typing-extensions==4.14.1
typing-inspection==0.4.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==15.0.1

# ---- Dev / docs (originally in ML list) ----
myst-parser==2.0.0
sphinx==7.1.2
sphinx-intl==2.0.1
sphinx-autodoc-typehints==1.12.0
pre-commit==2.15.0